wasuk_chatbot_new/requirements.txt
Flask==2.3.3
openai==1.3.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0 